+++ cookbooks/core/recipes/default.rb
@@ -20,7 +20,7 @@
 
 
 class Chef::Recipe
-  include Coderebels::Chefbox::Box
+  include Coderebels::Chefbox::Shell
   include Coderebels::Chefbox::Digest
 end